+++ core/src/main/java/com/ericsson/bss/cassandra/ecchronos/core/repair/RepairSchedulerImpl.java
@@ -221,8 +221,7 @@ public final class RepairSchedulerImpl implements RepairScheduler, Closeable
                 .withRepairPolices(myRepairPolicies)
                 .withRepairHistory(myRepairHistory)
                 .build();
-
-        job.runnable();
+        job.refreshState();
 
         return job;
     }