+++ core/src/main/java/com/ericsson/bss/cassandra/ecchronos/core/repair/TableRepairJob.java
@@ -202,7 +202,7 @@ public class TableRepairJob extends ScheduledJob
                         .withRepairHistory(myRepairHistory)
                         .withJobId(getId());
 
-                taskList.add(builder.build(getRealPriority()));
+                taskList.add(builder.build(getRealPriority(replicaRepairGroup.getLastCompletedAt())));
             }
 
             return taskList.iterator();
@@ -266,6 +266,15 @@ public class TableRepairJob extends ScheduledJob
     @Override
     public boolean runnable()
     {
+        return myRepairState.getSnapshot().canRepair() && super.runnable();
+    }
+
+    /**
+     * Refresh the repair state.
+     */
+    @Override
+    public void refreshState()
+    {
         try
         {
             myRepairState.update();
@@ -274,8 +283,6 @@ public class TableRepairJob extends ScheduledJob
         {
             LOG.warn("Unable to check repair history, {}", this, e);
         }
-
-        return myRepairState.getSnapshot().canRepair() && super.runnable();
     }
 
      /**