+++ core/src/main/java/com/ericsson/bss/cassandra/ecchronos/core/scheduling/ScheduledJobQueue.java
@@ -118,6 +118,7 @@ public class ScheduledJobQueue implements Iterable<ScheduledJob>
     @Override
     public final synchronized Iterator<ScheduledJob> iterator()
     {
+        myJobQueues.values().forEach(q -> q.forEach(ScheduledJob::refreshState));
         Iterator<ScheduledJob> baseIterator = new ManyToOneIterator<>(myJobQueues.values(), myComparator);
 
         return new RunnableJobIterator(baseIterator);